Remove insecure handling of verifier certificate
Disabling SSL certificate on the Persona verifier means that
an attacker could perform a trivial MITM attack (any SSL cert will
do) and take over authentication for users of a site.

This patch forces full SSL validation regardless of whether or not
an SSL bundle path has been configured in.